remote-playback/remote-video-playback-manual.html

<!DOCTYPE html>
<html>
  <title>Test if video is playing on remote device.</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/common/media.js"></script>
  <style>
    button {
        padding: 2em;
    }
  </style>
  <body>
    <div id="pick-device">
      <p>
        Click the button below to prompt for a remote playback device and select
        one!
      </p>
      <p>
        <button id="prompt-button">Pick device</button>
      </p>
    </div>
    <video src="/media/green-at-15.mp4" id="video"></video>
    <div id="evaluate" style="display: none">
      <p>Does the video play back on the remote device?</p>
      <p>
        <button id="yes">Yes</button>
      </p>
      <p>
        <button id="no">No</button>
      </p>
    </div>
  </body>
  <script>
    var v = document.getElementById("video");

    setup({
      explicit_timeout: true,
    });
    async_test((t) => {
      var button = document.getElementById("prompt-button");
      button.onclick = t.step_func(() => {
        promise_test(() => {
          return v.remote.prompt().then(() => {
            v.play();
            document.getElementById("evaluate").style.display = "block";
          });
        }, "Prompt resolves");
      });

      function evaluate(success) {
        assert_true(success, "Video plays on remote device");
      }
      var yesButton = document.getElementById("yes");
      var noButton = document.getElementById("no");
      yesButton.onclick = t.step_func_done(function () {
        evaluate(true);
      });
      noButton.onclick = t.step_func_done(function () {
        evaluate(false);
      });
    }, "Test if video is playing on remote device.");
  </script>
</html>
